fix(section-audio): duracion total del player (numeric pg -> string)
La columna duration_sec (numeric) llegaba como string desde postgres.js,
y fmtTime(string) caia en !Number.isFinite pintando "0:00". Coacciona en
SectionPlayer.paint (resolveDuration: Number + fallback a audio.duration
finito) y castea duration_sec::float8 en el endpoint para cumplir el
contrato number|null. Cubierto con 3 tests de regresion.

+++ api/songs/[id]/section-audio.js
@@ -14,7 +14,7 @@ const VOICE_SCOPES = ['soprano', 'contralto', 'tenor', 'bass', 'lead', 'backing'
 async function getSectionAudio(_req, res, songId) {
   const rows = await sql`
     SELECT id, section_index AS "sectionIndex", voice_scope AS "voiceScope",
-           storage_key AS "storageKey", label, duration_sec AS "durationSec"
+           storage_key AS "storageKey", label, duration_sec::float8 AS "durationSec"
     FROM song_section_audio
     WHERE song_id = ${songId}
     ORDER BY section_index, voice_scope
